+++ android/app/src/main/java/com/holabendez/tvshowwatchlist/MainActivity.java
@@ -1,0 +1,5 @@
+package com.holabendez.tvshowwatchlist;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}